+++ common-custom-user-data-maven-extension/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 - Change the type of the `log` variable bound for custom Groovy scripts to `org.slf4j.Logger`. This change may require changes in custom Groovy scripts.
 - Add custom value and search link for GitHub Actions run ID.
+- Redact any `user:password@` portion of the 'Git repository' custom value.
 
 ## [1.8.1] - 2021-10-13
 - Avoid use of `org.apache.maven.monitor.logging.DefaultLog` since this type has been renamed in Maven 4 (#165)